[e2e 95] use log dimension

+++ xinnovation/train.sh
@@ -4,8 +4,11 @@
 export PYTHONPATH=$PYTHONPATH:$(pwd)
 
 config_file=${1:-"./xinnovation/examples/detector4D/sparse4dv3_temporal_base.py"}
+delete_checkpoints=${2:-1}
 
-rm -rf ./xinnovation_visualize_intermediate_results 
-rm -rf ./xinnovation_checkpoints
+rm -rf ./xinnovation_visualize_intermediate_results
+if [ $delete_checkpoints -eq 1 ]; then
+    rm -rf ./xinnovation_checkpoints
+fi
 # 运行训练脚本
 python ./xinnovation/apis/train.py --config_file $config_file
